Moved RoundedToolbarShape to shapes directory.
