+++ android/app/src/main/java/com/koleff/kare_android/ui/compose/scaffolds/ExerciseDetailsScreenScaffold.kt
@@ -1,7 +1,6 @@
 package com.koleff.kare_android.ui.compose.scaffolds
 
 import ExerciseDetailsToolbar
-import RoundedToolbarShape
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -19,6 +18,7 @@ import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.koleff.kare_android.ui.compose.navigation.ExerciseDetailsBottomNavigationBar
+import com.koleff.kare_android.ui.compose.navigation.shapes.RoundedToolbarShape
 
 @Composable
 fun ExerciseDetailsScreenScaffold(
@@ -44,7 +44,7 @@ fun ExerciseDetailsScreenScaffold(
                     )
                     .border(
                         border = BorderStroke(2.dp, color = Color.White),
-                        shape = RoundedToolbarShape() //TODO: new shape without top part...
+                        shape = RoundedToolbarShape(hasTopOutline = false)
                     ),
                 isNavigationInProgress = isNavigationInProgress,
                 navController = navController,
